hybrid_ai_assistant/requirements.txt
langgraph
langchain
langchain-openai
langchain-community
ollama
openai
tavily-python
docker
pydantic
flask
python-dotenv
